py-consumer/requirements.txt
click==8.1.3
Flask==2.2.3
flask-sock==0.6.0
gunicorn==20.1.0
h11==0.14.0
itsdangerous==2.1.2
Jinja2==3.1.2
kafka-python==2.0.2
MarkupSafe==2.1.2
python-dotenv==1.0.0
simple-websocket==0.9.0
Werkzeug==2.2.3
wsproto==1.2.0
